Remove Home buttons and add them as items in the Drawer
