+++ app/src/main/java/com/example/msp_app/features/home/components/homefootersection/HomeFooterSection.kt
@@ -26,15 +26,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
 import com.example.msp_app.core.utils.Constants.APP_VERSION
 import com.example.msp_app.core.utils.ResultState
 import com.example.msp_app.data.api.ApiProvider
 import com.example.msp_app.data.models.payment.Payment
 import com.example.msp_app.data.models.visit.Visit
-import com.example.msp_app.features.productsInventory.viewmodels.ProductsInventoryViewModel
-import com.example.msp_app.navigation.Screen
 
 @Composable
 fun HomeFooterSection(
@@ -51,12 +47,9 @@ fun HomeFooterSection(
     onLogout: () -> Unit,
     onInitWeek: () -> Unit,
     modifier: Modifier = Modifier,
-    navController: NavController
 ) {
     val showDialogInitWeek = remember { mutableStateOf(false) }
     val baseURL = ApiProvider.baseURL.collectAsState()
-    val viewModel: ProductsInventoryViewModel = viewModel()
-    val productState = viewModel.productInventoryState.collectAsState().value
 
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
@@ -155,28 +148,6 @@ fun HomeFooterSection(
             is ResultState.Success -> Text("Ventas descargadas: ${syncSalesState.data.size}")
             is ResultState.Error -> Text("Error: ${syncSalesState.message}")
         }
-
-        Button(
-            onClick = { navController.navigate(Screen.ProductsCatalog.route) },
-            modifier = Modifier.fillMaxWidth(0.92f)
-        ) {
-            Text("CATALOGO DE PRODUCTOS", color = Color.White)
-        }
-
-        Button(
-            onClick = { viewModel.fetchRemoteInventory() },
-            modifier = Modifier.fillMaxWidth(0.92f)
-        ) {
-            Text("ACTUALIZAR CATALOGO", color = Color.White)
-        }
-        if (productState is ResultState.Loading) {
-            CircularProgressIndicator(
-                modifier = Modifier
-                    .padding(8.dp)
-                    .align(Alignment.CenterHorizontally)
-            )
-        }
-
         Button(
             onClick = {
                 onSyncPendingVisits()
